# openapi.yaml
openapi: 3.1.0
...
info:
  title: Nidan API
  version: 1.0.0
  description: |
    A bias-aware virtual patient simulator. Trainees interview an LLM-driven
    patient, order examinations and investigations, submit a diagnosis, and
    receive feedback on **how they reasoned**.

    Two guarantees shape this API and must not be broken by any change:

    - **No assessment state is exposed during a consultation** (`PRD` P2).
      There is no endpoint returning coverage, question counts, topic hints or
      readiness. Contract tests CT-1 and CT-2 assert their absence.
    - **No bias vocabulary appears in any response** (`PRD` P1). Not "bias",
      "anchoring", "premature closure" or "confirmation bias". Contract test
      CT-5 asserts this over rendered output.
servers:
  - url: https://api.nidan.app/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /me:
    post:
      tags: [identity]
      summary: Create or claim a profile
      description: >
        Idempotent. Pass `anonymous_id` to claim a trial session taken before
        signup; the window is 30 days (`PRD` FR-2).
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileCreate' }
      responses:
        # 201 on creation, 200 when the profile already existed. Both are
        # success — the guarantee is idempotency, not a fixed status — and the
        # difference tells a client whether onboarding still needs showing.
        '201': { $ref: '#/components/responses/Profile' }
        '200': { $ref: '#/components/responses/Profile' }
        '401': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
    get:
      tags: [identity]
      summary: The authenticated user's profile
      responses:
        '200': { $ref: '#/components/responses/Profile' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
    patch:
      tags: [identity]
      summary: Update profile fields
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdate' }
      responses:
        '200': { $ref: '#/components/responses/Profile' }
        '422': { $ref: '#/components/responses/Error' }
        '501': { $ref: '#/components/responses/NotImplemented' }
    delete:
      tags: [identity]
      summary: Delete the account
      description: >
        Requires explicit confirmation. Session results and the research
        pseudonym are retained so published analyses stay reproducible; the
        deletion screen must say so (`DATA_MODEL` §10.2).
      responses:
        '204': { description: Deleted }
        '422': { $ref: '#/components/responses/Error' }
        '501': { $ref: '#/components/responses/NotImplemented' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase-issued JWT, verified against JWKS (`ADR-0002`).
  responses:
    NotImplemented:
      description: >
        Not implemented. Every operation returns this until T-030; the contract
        exists first so the frontend can be built against it (sync point S-1).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Profile:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Profile' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: >
                Stable identifier. Once shipped a code is never renamed or
                repurposed — clients branch on it.
              enum:
                - unauthenticated
                - token_expired
                # 503, not 401. Returned when the key server cannot be reached,
                # so the token CANNOT be checked — which is not the same as a
                # token that failed a check. A 401 here would send every client
                # into a refresh loop against an outage (T-014).
                - service_unavailable
                - forbidden
                - pro_required
                - monthly_limit_reached
                - not_found
                - feedback_not_ready
                - session_already_active
                - session_concluded
                - session_not_concluded
                - trial_already_used
                - no_approving_review
                - validation_failed
                - question_too_long
                - diagnosis_empty
                - diagnosis_too_long
                - patient_data_detected
                - keyword_overlap
                - idempotency_key_reused
                - confirmation_required
                - invalid_timezone
                - trial_expired
                - unknown_key
                - question_limit_reached
                - rate_limited
                - patient_unavailable
                - no_cases_available
                - internal_error
            message: { type: string }
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
              description: Quote this to support; it correlates to the server logs.
    Profile:
      type: object
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string, nullable: true }
        professional_role:
          type: string
          nullable: true
          enum: [medical_student, intern, resident, physician, other]
        year_of_training: { type: integer, minimum: 1, maximum: 10, nullable: true }
        country: { type: string, minLength: 2, maxLength: 2, nullable: true }
        timezone: { type: string, example: Asia/Kolkata }
        subscription_tier: { type: string, enum: [free, pro] }
        consent_research: { type: boolean }
        onboarded_at: { type: string, format: date-time, nullable: true }
        sessions_remaining_this_month:
          type: integer
          nullable: true
          description: >
            Free tier only. Shown on the dashboard, NEVER during a consultation
            (`PRD` P2).
    ProfileCreate:
      type: object
      properties:
        anonymous_id:
          type: string
          nullable: true
          description: Claims a trial session taken before signup.
        timezone: { type: string }
    ProfileUpdate:
      type: object
      properties:
        display_name: { type: string }
        professional_role:
          type: string
          enum: [medical_student, intern, resident, physician, other]
        year_of_training: { type: integer, minimum: 1, maximum: 10 }
        country: { type: string, minLength: 2, maxLength: 2 }
        timezone: { type: string }
        consent_research: { type: boolean }
...